fix check bug
